# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreatePermissionRequest:
    properties:
      action:
        type: string
      description:
        type: string
      name:
        type: string
      scopeId:
        description: 权限域ID
        type: string
    required:
    - action
    - name
    type: object
...
  models.PermissionResponse:
    properties:
      action:
        type: string
      createTime:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      scope:
        allOf:
        - $ref: '#/definitions/models.ScopeResponse'
        description: 所属权限域
      updateTime:
        type: string
    type: object
...
  models.RoleResponse:
    properties:
      createTime:
        type: string
      description:
        type: string
      id:
        type: string
      inheritedBy:
        description: 被哪些角色继承
        items:
          $ref: '#/definitions/models.RoleResponse'
        type: array
      inheritsFrom:
        description: 继承的父角色
        items:
          $ref: '#/definitions/models.RoleResponse'
        type: array
      name:
        type: string
      permissions:
        description: 角色拥有的权限
        items:
          $ref: '#/definitions/models.PermissionResponse'
        type: array
      updateTime:
        type: string
    type: object
...
  models.UpdatePermissionRequest:
    properties:
      action:
        type: string
      description:
        type: string
      name:
        type: string
      scopeId:
        description: 权限域ID
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个基于Go和Gin框架的后端API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Backend API
  version: "1.0"
paths:
...
  /rbac/permissions:
    get:
      consumes:
      - application/json
      description: 根据分页参数获取权限列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: asc
        description: 排序方式
        in: query
        name: order
        type: string
      - default: id
        description: 排序字段
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/models.PermissionResponse'
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 分页获取权限列表
      tags:
      - rbac-permissions
    post:
      consumes:
      - application/json
      description: 创建新的权限
      parameters:
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/models.PermissionResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 创建权限
      tags:
      - rbac-permissions
  /rbac/permissions/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除权限
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 删除权限
      tags:
      - rbac-permissions
    get:
      consumes:
      - application/json
      description: 根据权限ID获取权限详细信息
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/models.PermissionResponse'
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 根据ID获取权限
      tags:
      - rbac-permissions
    put:
      consumes:
      - application/json
      description: 根据ID更新权限信息
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/models.PermissionResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 更新权限
      tags:
      - rbac-permissions
  /rbac/permissions/all:
    get:
      consumes:
      - application/json
      description: 获取系统中所有权限的列表（不分页）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  $ref: '#/definitions/models.PermissionResponse'
                type: array
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 获取所有权限
      tags:
      - rbac-permissions
  /rbac/roles:
    get:
      consumes:
      - application/json
      description: 根据分页参数获取角色列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: asc
        description: 排序方式
        in: query
        name: order
        type: string
      - default: id
        description: 排序字段
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/models.RoleResponse'
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 分页获取角色列表
      tags:
      - rbac-roles
    post:
      consumes:
      - application/json
      description: 创建新的角色
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/models.RoleResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 创建角色
      tags:
      - rbac-roles
  /rbac/roles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 删除角色
      tags:
      - rbac-roles
    get:
      consumes:
      - application/json
      description: 根据角色ID获取角色详细信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/models.RoleResponse'
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 根据ID获取角色
      tags:
      - rbac-roles
    put:
      consumes:
      - application/json
      description: 根据ID更新角色信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/models.RoleResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 更新角色
      tags:
      - rbac-roles
  /rbac/roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: 为角色分配权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID列表
        in: body
        name: permissions
        required: true
        schema:
          $ref: '#/definitions/models.AssignRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 分配角色权限
      tags:
      - rbac-roles
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"